Break measure<> out into own file
Doing a little cleanup on the sqltoaster binary. Just moving the
measure<> templated execution measurement class into its own header
file.
