+++ sqltoaster/main.cc
@@ -1,29 +1,19 @@
-// A demonstration of the sqltoast library API
+/*
+ * Use and distribution licensed under the Apache license version 2.
+ *
+ * See the COPYING file in the root project directory for full text.
+ */
 
 #include <string.h>
 
 #include <iostream>
-#include <chrono>
 
 #include <sqltoast/sqltoast.h>
 
+#include "measure.h"
 #include "printer.h"
 
 using namespace std;
-
-template<typename TimeT = std::chrono::milliseconds>
-struct measure
-{
-    template<typename F, typename ...Args>
-    static typename TimeT::rep execution(F&& func, Args&&... args)
-    {
-        auto start = std::chrono::steady_clock::now();
-        std::forward<decltype(func)>(func)(std::forward<Args>(args)...);
-        auto duration = std::chrono::duration_cast<TimeT>
-                            (std::chrono::steady_clock::now() - start);
-        return duration.count();
-    }
-};
 
 struct parser {
     sqltoast::parse_options_t opts;
